refactor(login): move login logic to navbar button
rename login rote to login-failed and only show notification
also move css to be scoped for login page
