+++ src/router/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import Home from '@/views/Home.vue'
-import Login from '@/views/Login.vue'
+import LoginFailed from '@/views/LoginFailed.vue'
 import Reset from '@/views/Reset.vue'
 import Dashboard from '@/views/private/Dashboard.vue'
 import Edit from '@/views/private/Edit.vue'
@@ -68,10 +68,10 @@ const routes = [
     meta: { requiresAuth: true, title: 'Update Study' }
   },
   {
-    path: '/login',
-    name: 'login',
-    component: Login,
-    meta: { title: 'Login' }
+    path: '/login-failed',
+    name: 'login-failed',
+    component: LoginFailed,
+    meta: { title: 'Login Failed' }
   },
   {
     path: '/register',
