+++ src/styles/themes/_default.sass
@@ -3,4 +3,3 @@
 @import './default/components/_form'
 @import './default/components/_dashboard'
 @import './default/components/_side-nav'
-@import './default/components/_login'
