--- src/styles/themes/default/components/_login.sass
@@ -1,12 +1,0 @@
-.login
-    display: flex
-    justify-content: center
-    align-items: center
-.login-container
-    margin-top: 10rem
-    width: 30rem
-    border: solid 1px #d3d3d3
-    border-radius: 5px
-    padding: 4rem
-    form
-        margin-bottom: 2rem
